Update puzzle_board_detector.py
In case of curved surfaces, it makes sense to have a less strict collinearity constraint for collinear code points. Now, with curved=True, collinear code points are allowed to appear in the image at an angle of up to 22.5° to each other. With default value curved=False, the maximum angle is set to 10°.